requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.0
python-multipart==0.0.9
pillow==10.3.0
transformers==4.46.2
torch==2.2.2
accelerate==0.30.1
bitsandbytes==0.43.1
gradio==4.24.0
numpy==1.26.4
httpx==0.27.0
